Rename ArchitecturyBlockEntity to BlockEntityExtension
