+++ forge/src/main/resources/architectury.mixins.json
@@ -5,7 +5,7 @@
   "minVersion": "0.8",
   "client": [
   ],
-  "mixins": ["BiomeGenerationSettingsBuilderAccessor", "MixinArchitecturyBlockEntity", "MixinBlockEntity", "MobSpawnSettingsBuilderAccessor"],
+  "mixins": ["BiomeGenerationSettingsBuilderAccessor", "MixinBlockEntity", "MixinBlockEntityExtension", "MobSpawnSettingsBuilderAccessor"],
   "injectors": {
     "defaultRequire": 1
   }
